Sketch out an alternative parallel prefetch algo
This commit sketches out the alternative proposal for prefetching
sources in parallel. The trick used was to split up the `prefetch`
function for the sources into two separate `start_prefetch` and
`finish_prefetch` functions.

With those we can maintain a simple queue to ensure we never run more
than the specified number of jobs. That being said this implementation
has a deficiency in that it has head-of-the-line blocking problem (i.e.
if the earliest spawned pre-fetch blocks for a long time, and later ones
finish earlier, we fail to spawn more work.

I still think its an improvement of status quo, especially given how
little complexity it introduces.